+++ htdocs/luci-static/resources/view/mstpd/status.js
@@ -17,7 +17,7 @@ var callMSTPStatus = rpc.declare({
 });
 
 const mstpdBridgeParams = {
-	'force-protocol-version'     : _('Administrative protocol version'),
+	'force-protocol-version'     : _('Administrative protocol'),
 	'enabled'                    : _('Enabled', 'Bridge status'),
 	'bridge-id'                  : _('Bridge ID'),
 	'designated-root'            : _('Designated root'),
@@ -35,15 +35,16 @@ const mstpdBridgeParams = {
 };
 
 const mstpdBridgePortParams = {
+	'port'                       : _('Port'),
 	'enabled'                    : _('Enabled', 'Port status'),
 	'bpdu-filter-port'           : _('STP/RSTP processing'),
 	'bpdu-guard-port'            : _('BPDU guard'),
 	'bpdu-guard-error'           : _('BPDU guard error'),
-	'received-stp'               : _('Operative protocol version'),
-	'role'                       : _('Port role'),
-	'state'                      : _('Port state'),
-	'port-id'                    : _('Port identifier'),
-	'external-port-cost'         : _('Port path cost'),
+	'received-stp'               : _('Operative protocol'),
+	'role'                       : _('Role'),
+	'state'                      : _('State'),
+	'port-id'                    : _('Identifier'),
+	'external-port-cost'         : _('Path cost'),
 	'designated-root'            : _('Designated root'),
 	'designated-bridge'          : _('Designated bridge'),
 	'designated-port'            : _('Designated port'),
@@ -56,7 +57,9 @@ const mstpdBridgePortParams = {
 	'num-rx-bpdu'                : _('Rx BPDU'),
 	'num-rx-bpdu-filtered'       : _('Rx BPDU (filtered)'),
 	'num-tx-tcn'                 : _('Tx TCN'),
-	'num-rx-tcn'                 : _('Rx TCN')
+	'num-rx-tcn'                 : _('Rx TCN'),
+	'num-transition-fwd'         : _('Transitions to "Forwarding" state'),
+	'num-transition-blk'         : _('Transitions to "Blocking" state'),
 };
 
 function vcb_yn(br, v) {
@@ -81,7 +84,7 @@ function vcb_time(br, v) {
 	return _('%t (%d s)').format(v, v);
 }
 
-function mstpd_br_status_add(br, option, value_cb_func)
+function mstpd_br_param_add(br, option, value_cb_func)
 {
 	if (!mstpdBridgeParams[option] || (typeof br[option] == 'undefined'))
 		return '';
@@ -90,8 +93,8 @@ function mstpd_br_status_add(br, option, value_cb_func)
 		value_cb_func = function(br, v) { return v; }
 
 	return E('div', {}, [
-		E('span', { 'class': 'mstpd-br-param' }, mstpdBridgeParams[option]),
-		E('span', { 'class': 'mstpd-br-param-value' }, value_cb_func(br, br[option]))
+		E('span', { 'class': 'mstpd-param' }, mstpdBridgeParams[option]),
+		E('span', { 'class': 'mstpd-param-value' }, value_cb_func(br, br[option]))
 	]);
 }
 
@@ -106,35 +109,35 @@ var CBIBridgeStatus = form.DummyValue.extend({
 			);
 		}
 
-		items.push(mstpd_br_status_add(br, 'enabled', vcb_yn));
+		items.push(mstpd_br_param_add(br, 'enabled', vcb_yn));
 
-		items.push(mstpd_br_status_add(br, 'force-protocol-version', function(br, v) {
+		items.push(mstpd_br_param_add(br, 'force-protocol-version', function(br, v) {
 			return v.toUpperCase()
 		}));
 
-		items.push(mstpd_br_status_add(br, 'bridge-id'));
-		items.push(mstpd_br_status_add(br, 'designated-root'));
-		items.push(mstpd_br_status_add(br, 'regional-root'));
+		items.push(mstpd_br_param_add(br, 'bridge-id'));
+		items.push(mstpd_br_param_add(br, 'designated-root'));
+		items.push(mstpd_br_param_add(br, 'regional-root'));
 
-		items.push(mstpd_br_status_add(br, 'root-port', function(br, v) {
+		items.push(mstpd_br_param_add(br, 'root-port', function(br, v) {
 			return ((v == 'none') || (v == '')) ? _('none', 'Port name') : v
 		}));
 
-		items.push(mstpd_br_status_add(br, 'path-cost'));
-		items.push(mstpd_br_status_add(br, 'max-age', vcb_seconds));
-		items.push(mstpd_br_status_add(br, 'forward-delay', vcb_seconds));
-		items.push(mstpd_br_status_add(br, 'tx-hold-count'));
-		items.push(mstpd_br_status_add(br, 'hello-time', vcb_seconds));
-		items.push(mstpd_br_status_add(br, 'ageing-time', vcb_seconds));
-		items.push(mstpd_br_status_add(br, 'time-since-topology-change', vcb_time));
+		items.push(mstpd_br_param_add(br, 'path-cost'));
+		items.push(mstpd_br_param_add(br, 'max-age', vcb_seconds));
+		items.push(mstpd_br_param_add(br, 'forward-delay', vcb_seconds));
+		items.push(mstpd_br_param_add(br, 'tx-hold-count'));
+		items.push(mstpd_br_param_add(br, 'hello-time', vcb_seconds));
+		items.push(mstpd_br_param_add(br, 'ageing-time', vcb_seconds));
+		items.push(mstpd_br_param_add(br, 'time-since-topology-change', vcb_time));
 
-		items.push(mstpd_br_status_add(br, 'last-topology-change-port', function(br, v) {
+		items.push(mstpd_br_param_add(br, 'last-topology-change-port', function(br, v) {
 			return ((v == 'None') || (v == '')) ? _('none', 'Port name') : v
 		}));
 
-		items.push(mstpd_br_status_add(br, 'topology-change-count'));
+		items.push(mstpd_br_param_add(br, 'topology-change-count'));
 
-		return E('div', { 'class': 'mstpd-br-params' }, items);
+		return E('div', { 'class': 'mstpd-params mstpd-br-params' }, items);
 	},
 });
 
@@ -168,130 +171,261 @@ function vcb_num_or_dash(br, v) {
 	return parseInt(v) == 0 ? '–' : v;
 }
 
-function mstpd_brports_status_add(br, option, value_cb_func, bpdu_filter)
+function mstpd_br_port_param_add(br, port, option, value_cb_func, bpdu_filter)
 {
-	var row = [];
-
 	if (!mstpdBridgePortParams[option])
 		return null;
 
 	if (!value_cb_func)
 		value_cb_func = function(br, v) { return v };
 
-	row.push(mstpdBridgePortParams[option]);
-
-	br.ports.forEach(function(p) {
-		if (bpdu_filter) {
-			if (p['bpdu-filter-port'] == 'yes') {
-				row.push('–');
-				return;
-			}
+	if (bpdu_filter) {
+		if (port['bpdu-filter-port'] == 'yes') {
+			return '–';
 		}
+	}
 
-		row.push(value_cb_func(br, p[option]));
-	})
-
-	return row;
+	return value_cb_func(br, port[option]);
 }
+
+function mstpd_br_port_ex_param_add(br, port, option, value_cb_func, bpdu_filter)
+{
+	var value = mstpd_br_port_param_add(br, port, option, value_cb_func, bpdu_filter);
+
+	if (value === '')
+		return value;
+
+	return E('div', {}, [
+		E('span', { 'class': 'mstpd-param' }, mstpdBridgePortParams[option]),
+		E('span', { 'class': 'mstpd-param-value' }, value)
+	]);
+}
+
+var rowsExpanded = [];
 
 var CBIBridgePortsStatus = form.DummyValue.extend({
 	__init__: function() {
 		this.super('__init__', arguments);
 
 		this.recentData = null;
-		this.table = E('div', { 'class': 'table mstpd-brports-status' }, [
-			E('div', { 'class': 'tr table-titles' }, [
-				E('div', { 'class': 'th top center' }, '...')
+		this.table = E('table', { 'class': 'table mstpd-brports-status' }, [
+			E('tr', { 'class': 'tr table-titles' }, [
+				E('th', { 'class': 'th top left'  }, mstpdBridgePortParams['port']),
+				E('th', { 'class': 'th top left'  }, mstpdBridgePortParams['enabled']),
+				E('th', { 'class': 'th top left'  }, mstpdBridgePortParams['received-stp']),
+				E('th', { 'class': 'th top right' }, mstpdBridgePortParams['role']),
+				E('th', { 'class': 'th top left'  }, mstpdBridgePortParams['state']),
+				E('th', { 'class': 'th top right' }, mstpdBridgePortParams['port-id']),
+				E('th', { 'class': 'th top right' }, mstpdBridgePortParams['num-tx-bpdu']),
+				E('th', { 'class': 'th top right' }, mstpdBridgePortParams['num-rx-bpdu']),
 			]),
-			E('div', { 'class': 'tr placeholder' }, [
-				E('div', { 'class': 'td' },
+			E('tr', { 'class': 'tr placeholder' }, [
+				E('td', { 'class': 'td' },
 					E('em', { 'class': 'spinning' }, _('Collecting data...'))
 				)
 			])
 		]);
 	},
 
-	tableUpdateHeaders: function(br) {
-		var p;
-		var len = br.ports.length;
+	/** @private */
+	handleToggleRow: function(row, row_id) {
+		var e_img      = row.querySelector('img');
+		var e_extrow   = row.nextSibling;
 
-		if (!this.recentData || (len != recentData.ports.length))
-		{
-			var portHeaders = [];
-			portHeaders.push(E('div', { 'class': 'th middle left' }, _('Port')));
+		var do_expand = (e_extrow.style.display === 'none');
 
-			for (p = 0; p < len; p++)
-			{
-				portHeaders.push(
-					E('div', {
-						'class': 'th middle right',
-						'id': 'mstpd-brports-status-header-' + p
-					}, br.ports[p].port)
-				);
-			}
-
-			var header = E('div', { 'class': 'tr table-titles' }, portHeaders);
-			dom.content(this.table, header);
+		if (do_expand) {
+			row.classList.remove('mstpd-collapsed');
+			row.classList.add('mstpd-expanded');
 		}
-		else
-		{
-			for (p = 0; p < len; p++)
-			{
-				var header = this.table.querySelector(
-					'#mstpd-brports-status-header-' + p)
-
-				dom.content(header, br.ports[p].port);
-			}
+		else {
+			row.classList.remove('mstpd-expanded');
+			row.classList.add('mstpd-collapsed');
 		}
+
+		rowsExpanded[row_id] = do_expand;
+
+		e_extrow.style.display = do_expand ? 'table-row' : 'none';
+		e_img.src = this.getFoldingImage(do_expand);
 	},
 
-	tableUpdateData: function(br, shortMode) {
+	/** @private */
+	updateTable: function(table, data, placeholder) {
+		var target = isElem(table) ? table : document.querySelector(table);
+
+		if (!isElem(target))
+			return;
+
+		target.querySelectorAll(
+			'.tr.table-titles, .cbi-section-table-titles').forEach(L.bind(function(thead) {
+			var titles = [];
+
+			thead.querySelectorAll('.th').forEach(function(th) {
+				titles.push(th);
+			});
+
+			if (Array.isArray(data)) {
+				var n = 0, rows = target.querySelectorAll('.tr');
+
+				data.forEach(L.bind(function(row) {
+					var id = row[0];
+					var is_expanded = rowsExpanded[id] || false;
+					var trow = E('tr', {
+						'id': 'row-' + id + '-normal',
+						'class': 'tr ' + (is_expanded ? 'mstpd-expanded' : 'mstpd-collapsed'),
+						'click': L.bind(function(ev) {
+							this.handleToggleRow(ev.currentTarget, id);
+						}, this)
+					});
+
+					for (var i = 0; i < titles.length; i++) {
+						var text = (titles[i].innerText || '').trim();
+						var td = trow.appendChild(E('td', {
+							'class': titles[i].className,
+							'data-title': (text !== '') ? text : null
+						}, row[i + 1] || ''));
+
+						td.classList.remove('th');
+						td.classList.add('td');
+					}
+
+					var rowstyle = (n++ % 2) ? 2 : 1;
+
+					trow.classList.add('cbi-rowstyle-%d'.format(rowstyle));
+
+					if (rows[n])
+						target.replaceChild(trow, rows[n]);
+					else
+						target.appendChild(trow);
+
+					if (typeof row[titles.length + 1] !== 'undefined') {
+						/* Add expanded information row */
+						var trow = E('tr', {
+							'id': 'row-' + id + '-extended',
+							'class': 'tr mstpd-extrow',
+							'style': 'display: ' +
+								(is_expanded ? 'table-row' : 'none') + ';',
+						});
+
+						var td = trow.appendChild(E('td', {
+							'class': 'td', 'colspan': titles.length
+						}, row[titles.length + 1]));
+
+						trow.classList.add('cbi-rowstyle-%d'.format(rowstyle));
+
+						if (rows[n + 1])
+							target.replaceChild(trow, rows[n + 1]);
+						else
+							target.appendChild(trow);
+					}
+
+				}, this));
+
+				while (rows[++n])
+					target.removeChild(rows[n]);
+
+				if (placeholder && target.firstElementChild === target.lastElementChild) {
+					var trow = target.appendChild(
+						E('tr', { 'class': 'tr placeholder' }));
+
+					var td = trow.appendChild(
+						E('td', { 'class': 'center ' + titles[0].className }, placeholder));
+
+					td.classList.remove('th');
+					td.classList.add('td');
+				}
+			} else {
+				thead.parentNode.style.display = 'none';
+
+				thead.parentNode.querySelectorAll('.tr, .cbi-section-table-row').forEach(function(trow) {
+					if (trow !== thead) {
+						var n = 0;
+						trow.querySelectorAll('.th, .td').forEach(function(td) {
+							if (n < titles.length) {
+								var text = (titles[n++].innerText || '').trim();
+								if (text !== '')
+									td.setAttribute('data-title', text);
+							}
+						});
+					}
+				});
+
+				thead.parentNode.style.display = '';
+			}
+		}, this));
+	},
+
+	/** @private */
+	getFoldingImage: function(expanded) {
+		return L.resource('mstpd/details_' +
+			(expanded ? 'hide' : 'show') + '.svg');
+	},
+
+	tableUpdateData: function(br) {
 		var rows = [];
 
-		rows.push(mstpd_brports_status_add(br, 'enabled', vcb_yn));
-		rows.push(mstpd_brports_status_add(br, 'bpdu-filter-port', vcb_yn_inverted));
+		br.ports.forEach(function(p) {
+			var items = [];
 
-		rows.push(mstpd_brports_status_add(br, 'received-stp', function(br, v) {
-			if (br['force-protocol-version'] === 'stp') {
-				return E('span', { 'class': 'mstpd-badge mstpd-proto-ok' }, 'STP');
-			} else {
-				if (v === 'yes')
-					return E('span', { 'class': 'mstpd-badge mstpd-proto-changed' }, 'STP');
-				else
-					return E('span', { 'class': 'mstpd-badge mstpd-proto-ok' }, 'RSTP');
-			}
-		}, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'bpdu-filter-port', vcb_yn_inverted));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'designated-root', null, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'designated-bridge', null, true));
 
-		rows.push(mstpd_brports_status_add(br, 'role', vcb_port_role, true));
-		rows.push(mstpd_brports_status_add(br, 'state', vcb_port_state, true));
-		rows.push(mstpd_brports_status_add(br, 'port-id'));
-		rows.push(mstpd_brports_status_add(br, 'designated-root', null, true));
-		rows.push(mstpd_brports_status_add(br, 'designated-bridge', null, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'designated-port', function(br, v) {
+				if (v === '0.000')
+					return '–';
+				return v;
+			}, true));
 
-		rows.push(mstpd_brports_status_add(br, 'designated-port', function(br, v) {
-			if (v === '0.000')
-				return '–';
-			return v;
-		}, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'external-port-cost', null, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'admin-edge-port', vcb_yn, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'auto-edge-port', vcb_yn, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'oper-edge-port', vcb_yn, true)) ;
+			items.push(mstpd_br_port_ex_param_add(br, p, 'admin-point-to-point', vcb_yna, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'point-to-point', vcb_yn, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'bpdu-guard-port', vcb_yn));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'bpdu-guard-error', vcb_yn));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'num-rx-bpdu-filtered', vcb_num_or_dash));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'num-tx-tcn', vcb_num_or_dash, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'num-rx-tcn', vcb_num_or_dash, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'num-transition-fwd', vcb_num_or_dash, true));
+			items.push(mstpd_br_port_ex_param_add(br, p, 'num-transition-blk', vcb_num_or_dash, true));
 
-		if (!shortMode)
-		{
-			rows.push(mstpd_brports_status_add(br, 'external-port-cost', null, true));
-			rows.push(mstpd_brports_status_add(br, 'admin-edge-port', vcb_yn, true));
-			rows.push(mstpd_brports_status_add(br, 'auto-edge-port', vcb_yn, true));
-			rows.push(mstpd_brports_status_add(br, 'oper-edge-port', vcb_yn, true)) ;
-			rows.push(mstpd_brports_status_add(br, 'admin-point-to-point', vcb_yna, true));
-			rows.push(mstpd_brports_status_add(br, 'point-to-point', vcb_yn, true));
-			rows.push(mstpd_brports_status_add(br, 'bpdu-guard-port', vcb_yn));
-			rows.push(mstpd_brports_status_add(br, 'bpdu-guard-error', vcb_yn));
-			rows.push(mstpd_brports_status_add(br, 'num-tx-bpdu', vcb_num_or_dash, true));
-			rows.push(mstpd_brports_status_add(br, 'num-rx-bpdu', vcb_num_or_dash, true));
-			rows.push(mstpd_brports_status_add(br, 'num-rx-bpdu-filtered', vcb_num_or_dash));
-			rows.push(mstpd_brports_status_add(br, 'num-tx-tcn', vcb_num_or_dash, true));
-			rows.push(mstpd_brports_status_add(br, 'num-rx-tcn', vcb_num_or_dash, true));
-		}
+			var expanded_info = E('div', { 'class': 'mstpd-params mstpd-port-params' }, items);
 
-		cbi_update_table(this.table, rows,
+			var id = p['bridge'] + '-' + p['port'];
+
+			rows.push([
+				id,
+				E('div', { 'style': 'display: flex; flex-flow: row;' }, [
+					E('img', {
+						'class': 'mstpd-folding',
+						'style': 'width: 16px; margin-right: 8px;',
+						'src': this.getFoldingImage(rowsExpanded[id] || false)
+					}),
+					E('span', {}, mstpd_br_port_param_add(br, p, 'port')),
+				]),
+				mstpd_br_port_param_add(br, p, 'enabled', vcb_yn),
+				mstpd_br_port_param_add(br, p, 'received-stp', function(br, v) {
+					if (br['force-protocol-version'] === 'stp') {
+						return E('span', { 'class': 'mstpd-badge mstpd-proto-ok' }, 'STP');
+					} else {
+						if (v === 'yes')
+							return E('span', { 'class': 'mstpd-badge mstpd-proto-changed' }, 'STP');
+						else
+							return E('span', { 'class': 'mstpd-badge mstpd-proto-ok' }, 'RSTP');
+					}
+				}, true),
+				mstpd_br_port_param_add(br, p, 'role', vcb_port_role, true),
+				mstpd_br_port_param_add(br, p, 'state', vcb_port_state, true),
+				mstpd_br_port_param_add(br, p, 'port-id'),
+				mstpd_br_port_param_add(br, p, 'num-tx-bpdu', vcb_num_or_dash, true),
+				mstpd_br_port_param_add(br, p, 'num-rx-bpdu', vcb_num_or_dash, true),
+				expanded_info
+			]);
+		}.bind(this));
+
+		this.updateTable(this.table, rows, 
 			_('There is no available information'));
 	},
 
@@ -304,34 +438,10 @@ var CBIBridgePortsStatus = form.DummyValue.extend({
 				  'configured to be controlled by the STP/RSTP service.'));
 		}
 
-		this.shortMode = mstpd.getStatusShortMode(br['bridge']);
-
-		this.tableUpdateHeaders(br);
-		this.tableUpdateData(br, this.shortMode);
+		this.tableUpdateData(br);
 		this.recentData = br;
 
-		function btnTitle(shortMode) {
-			return shortMode
-				? ('↓ ' + _('More information') + ' ↓')
-				: ('↑ ' + _('Less information') + ' ↑');
-		}
-
-		return E('div', {}, [
-			E('div', { 'class': 'table-wrapper' }, this.table),
-			E('div', { 'class': 'cbi-page-actions control-group' }, [
-				E('input', {
-					'class': 'cbi-button',
-					'type': 'button',
-					'value': btnTitle(this.shortMode),
-					'click': L.bind(function(ev) {
-						this.shortMode = !mstpd.getStatusShortMode(this.recentData['bridge']);
-						mstpd.setStatusShortMode(this.recentData['bridge'], this.shortMode);
-						this.tableUpdateData(this.recentData, this.shortMode);
-						ev.currentTarget.value = btnTitle(this.shortMode);
-					}, this)
-				})
-			])
-		]);
+		return E('div', { 'class': 'table-wrapper' }, this.table);
 	}
 });
 
@@ -423,7 +533,6 @@ return L.view.extend({
 			var ss = o.subsection;
 			o = ss.option(CBIBridgePortsStatus, '_brportsinfo');
 			o.cfgvalue = L.bind(function(br) { return br; }, o, br);
-			o.shortMode = true;
 		}
 
 		return m.render();